+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Changelog
 
 ### Fixed
 - Fixed a bug which prevented the `oauthProviderOptions` config from being set from a config file for some providers and from the plugin’s stored settings for other providers.
+- Fixed a bug where videos wouldn’t automatically start to play when clicking on the play button.
 
 ## 2.0.1 - 2018-05-26